requirements.txt
fsm @ git+https://github.com/antonio-projects-studio/fsm.git@2cdf4f9296e7af4321f56f55167c1d0115679e9a
python-frontmatter==1.1.0
PyYAML==5.1
setuptools==70.0.0
